local-requirements.txt
pylint
pytest
pytest-flask